Rename the package and test file names to something shorter.
- With different package names, two files can share the same name without introducing file naming conflicts.
